Step 6: Whitelist Type Matching Rules
Sometimes, we do need to specify some deserialization hints in our
payloads. However, these should be controlled against a whitelist.

With Jackson, this can be achieved with @JsonTypeInfo and
@JsonSubTypes, and other libraries have comparable facilities. The
main difference here is that the deserializer has a configured set
of logical values to pick from for selecting what Java class to
deserialize into.